cross_trust_domain_demo/requirements.txt
# Requirements for Cross-Trust-Domain Identity Verification Demo
aztp-client>=0.1.11
python-dotenv>=0.19.0
termcolor>=1.1.0 